Added Contact Repository interface,  Contact Repository implementation in Application

+++ hw41/src/main/java/com/chilik1020/hw41/views/list/OnRecyclerViewItemClickListener.kt
@@ -1,5 +1,5 @@
 package com.chilik1020.hw41.views.list
 
 interface OnRecyclerViewItemClickListener {
-    fun onClick(id: Long)
+    fun onClick(id: Int)
 }
